ai-service/requirements.txt
uvicorn[standard]==0.34.2
openai==1.3.0
pydantic==2.11.2
pydantic-settings==2.0.3
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.1.0
google-generativeai==0.8.5
fastapi==0.110.0
groq==0.26.0
pymongo==4.6.0
